Relocate to the .uuid package, per @ikwattro feedback.
https://github.com/graphaware/neo4j-uuid/pull/31
