+++ src/main/java/com/graphaware/module/uuid/GraphDatabaseServiceAware.java
@@ -13,7 +13,7 @@
  * the GNU General Public License along with this program.  If not, see
  * <http://www.gnu.org/licenses/>.
  */
-package com.graphaware.module;
+package com.graphaware.module.uuid;
 
 import org.neo4j.graphdb.GraphDatabaseService;
